Extract utils and filter languages based on completeness
