+++ package.json
@@ -12,7 +12,7 @@
     "init": "node -r dotenv/config tasks/download_translations.js",
     "build": "npm-run-all -p build:*",
     "build:static": "mkdir -p dist && cp -r source/src/fonts source/src/img source/src/video dist",
-    "build:template": "node tasks/render_templates.js"
+    "build:vtt": "node tasks/render_vtt.js"
   },
   "keywords": [
     "BTCPayServer"
